[Refactor] LocalStorage key값 상수화
