+++ src/constants/storageKey.js
@@ -1,0 +1,1 @@
+export const STORAGE_KEY = "todoList"
