+++ src/utils/getHashFilter.js
